requirements.txt
flask
asyncio~=3.4.3
requests~=2.32.3
os
msal
webbrowser
python-dotenv~=1.0.1
groq~=0.11.0
pandas~=2.2.3
tabulate~=0.9.0
SQLAlchemy~=2.0.36